+++ src/tests/minlength.test.ts
@@ -7,10 +7,43 @@ describe('min length', () => {
     })
 
     const numbers = [1, 2, 3]
-    const id = '75JILToVsGerOADWmHlY38xvbaNZKQ9wdFS0B6kcMEtnRpgizhjU42qT1cd0dL'
+    const id = '86Rf07xd4zBmiJXQG6otHEbew02c3PWsUOLZxADhCpKj7aVFv9I8RquYrNlSTM'
 
     expect(sqids.encode(numbers)).toBe(id)
     expect(sqids.decode(id)).toEqual(numbers)
+  })
+
+  it('incremental', () => {
+    const numbers = [1, 2, 3]
+
+    const map = {
+      6: '86Rf07',
+      7: '86Rf07x',
+      8: '86Rf07xd',
+      9: '86Rf07xd4',
+      10: '86Rf07xd4z',
+      11: '86Rf07xd4zB',
+      12: '86Rf07xd4zBm',
+      13: '86Rf07xd4zBmi',
+      [defaultOptions.alphabet.length + 0]:
+        '86Rf07xd4zBmiJXQG6otHEbew02c3PWsUOLZxADhCpKj7aVFv9I8RquYrNlSTM',
+      [defaultOptions.alphabet.length + 1]:
+        '86Rf07xd4zBmiJXQG6otHEbew02c3PWsUOLZxADhCpKj7aVFv9I8RquYrNlSTMy',
+      [defaultOptions.alphabet.length + 2]:
+        '86Rf07xd4zBmiJXQG6otHEbew02c3PWsUOLZxADhCpKj7aVFv9I8RquYrNlSTMyf',
+      [defaultOptions.alphabet.length + 3]:
+        '86Rf07xd4zBmiJXQG6otHEbew02c3PWsUOLZxADhCpKj7aVFv9I8RquYrNlSTMyf1',
+    }
+
+    for (const [minLength, id] of Object.entries(map)) {
+      const sqids = new Sqids({
+        minLength: Number(minLength),
+      })
+
+      expect(sqids.encode(numbers)).toBe(id)
+      expect(sqids.encode(numbers)).toHaveLength(Number(minLength))
+      expect(sqids.decode(id)).toEqual(numbers)
+    }
   })
 
   it('incremental numbers', () => {
@@ -19,16 +52,16 @@ describe('min length', () => {
     })
 
     const ids = {
-      jf26PLNeO5WbJDUV7FmMtlGXps3CoqkHnZ8cYd19yIiTAQuvKSExzhrRghBlwf: [0, 0],
-      vQLUq7zWXC6k9cNOtgJ2ZK8rbxuipBFAS10yTdYeRa3ojHwGnmMV4PDhESI2jL: [0, 1],
-      YhcpVK3COXbifmnZoLuxWgBQwtjsSaDGAdr0ReTHM16yI9vU8JNzlFq5Eu2oPp: [0, 2],
-      OTkn9daFgDZX6LbmfxI83RSKetJu0APihlsrYoz5pvQw7GyWHEUcN2jBqd4kJ9: [0, 3],
-      h2cV5eLNYj1x4ToZpfM90UlgHBOKikQFvnW36AC8zrmuJ7XdRytIGPawqYEbBe: [0, 4],
-      '7Mf0HeUNkpsZOTvmcj836P9EWKaACBubInFJtwXR2DSzgYGhQV5i4lLxoT1qdU': [0, 5],
-      APVSD1ZIY4WGBK75xktMfTev8qsCJw6oyH2j3OnLcXRlhziUmpbuNEar05QCsI: [0, 6],
-      P0LUhnlT76rsWSofOeyRGQZv1cC5qu3dtaJYNEXwk8Vpx92bKiHIz4MgmiDOF7: [0, 7],
-      xAhypZMXYIGCL4uW0te6lsFHaPc3SiD1TBgw5O7bvodzjqUn89JQRfk2Nvm4JI: [0, 8],
-      '94dRPIZ6irlXWvTbKywFuAhBoECQOVMjDJp53s2xeqaSzHY8nc17tmkLGwfGNl': [0, 9],
+      SvIzsqYMyQwI3GWgJAe17URxX8V924Co0DaTZLtFjHriEn5bPhcSkfmvOslpBu: [0, 0],
+      n3qafPOLKdfHpuNw3M61r95svbeJGk7aAEgYn4WlSjXURmF8IDqZBy0CT2VxQc: [0, 1],
+      tryFJbWcFMiYPg8sASm51uIV93GXTnvRzyfLleh06CpodJD42B7OraKtkQNxUZ: [0, 2],
+      eg6ql0A3XmvPoCzMlB6DraNGcWSIy5VR8iYup2Qk4tjZFKe1hbwfgHdUTsnLqE: [0, 3],
+      rSCFlp0rB2inEljaRdxKt7FkIbODSf8wYgTsZM1HL9JzN35cyoqueUvVWCm4hX: [0, 4],
+      sR8xjC8WQkOwo74PnglH1YFdTI0eaf56RGVSitzbjuZ3shNUXBrqLxEJyAmKv2: [0, 5],
+      uY2MYFqCLpgx5XQcjdtZK286AwWV7IBGEfuS9yTmbJvkzoUPeYRHr4iDs3naN0: [0, 6],
+      '74dID7X28VLQhBlnGmjZrec5wTA1fqpWtK4YkaoEIM9SRNiC3gUJH0OFvsPDdy': [0, 7],
+      '30WXpesPhgKiEI5RHTY7xbB1GnytJvXOl2p0AcUjdF6waZDo9Qk8VLzMuWrqCS': [0, 8],
+      moxr3HqLAK0GsTND6jowfZz3SUx7cQ8aC54Pl1RbIvFXmEJuBMYVeW9yrdOtin: [0, 9],
     }
 
     for (const [id, numbers] of Object.entries(ids)) {
@@ -38,17 +71,15 @@ describe('min length', () => {
   })
 
   it('min lengths', () => {
-    const s = new Sqids()
-
     for (const minLength of [0, 1, 5, 10, defaultOptions.alphabet.length]) {
       for (const numbers of [
-        [s.minValue()],
+        [0],
         [0, 0, 0, 0, 0],
         [1, 2, 3, 4, 5, 6, 7, 8, 9, 10],
         [100, 200, 300],
         [1_000, 2_000, 3_000],
         [1_000_000],
-        [s.maxValue()],
+        [Number.MAX_SAFE_INTEGER],
       ]) {
         const sqids = new Sqids({
           minLength,
@@ -62,23 +93,21 @@ describe('min length', () => {
   })
 
   it('out-of-range invalid min length', () => {
-    const s = new Sqids()
-    const err = `Minimum length has to be between ${s.minValue()} and ${
-      defaultOptions.alphabet.length
-    }`
+    const minLengthLimit = 255
+    const minLengthError = `Minimum length has to be between 0 and ${minLengthLimit}`
 
     expect(
       () =>
         new Sqids({
           minLength: -1,
         }),
-    ).toThrow(err)
+    ).toThrow(minLengthError)
 
     expect(
       () =>
         new Sqids({
-          minLength: defaultOptions.alphabet.length + 1,
+          minLength: minLengthLimit + 1,
         }),
-    ).toThrow(err)
+    ).toThrow(minLengthError)
   })
 })